Crate gossip_kv

Modules

lattices

membership

model

server

util

Structs

Key

A key of an entry in the key-value store.

Enums

ClientRequest

A request from a client to the key-value store.

ClientResponse

GossipMessage

KeyParseError

Error that can occur when parsing a key from a string.

Namespace

The namespace of the key of an entry in the key-value store.

Type Aliases

RowKey

The key of a row in a table in the key-value store.

TableName

The name of a table in the key-value store.