gossip_kv

Module model

Functions

delete_row

TableMap element to delete a row from an existing TableMap.

upsert_row

TableMap element to upsert a row in an existing TableMap.

Type Aliases

Clock

Timestamps used in the model.

NamespaceMap

Namespaces

RowKey

Primary key for entries in a table.

RowValue

Value stored in a table. Modelled as a timestamped set of strings.

Table

A map from row keys to values in a table.

TableMap

A map from table names to tables.

TableName

Name of a table in the data store.